+++ context/app/static/js/components/entity-search/Search/hooks.js
@@ -1,21 +1,17 @@
 import { useState, useCallback, useContext, useMemo } from 'react';
 import { useSearchkitVariables } from '@searchkit/client';
 import { CustomQuery } from '@searchkit/sdk';
-import useDeepCompareEffect from 'use-deep-compare-effect';
 
 import useSearchkitSDK from 'js/components/entity-search/searchkit-modifications/useSearchkitSDK';
 import { AppContext } from 'js/components/Providers';
 import { getAuthHeader } from 'js/helpers/functions';
 import { useStore } from 'js/components/entity-search/SearchWrapper/store';
 import { createSearchkitFacet } from 'js/components/entity-search/SearchWrapper/utils';
-import { fetchSearchData } from 'js/hooks/useSearchData';
 
 import { buildSortPairs, getRangeProps } from './utils';
 
-function useAllResultsUUIDs() {
+function useQueryBody() {
   const [queryBody, setQueryBody] = useState({});
-  const [allResultsUUIDs, setAllResultsUUIDS] = useState([]);
-  const { elasticsearchEndpoint, groupsToken } = useContext(AppContext);
 
   const setQueryBodyAndReturnBody = useCallback(
     (body) => {
@@ -25,21 +21,7 @@ function useAllResultsUUIDs() {
     [setQueryBody],
   );
 
-  useDeepCompareEffect(() => {
-    async function getAndSetAllUUIDs() {
-      const { query, post_filter } = queryBody;
-      const allResults = await fetchSearchData(
-        { query, post_filter, _source: false, size: 10000 },
-        elasticsearchEndpoint,
-        groupsToken,
-      );
-      // eslint-disable-next-line no-underscore-dangle
-      setAllResultsUUIDS(allResults.hits.hits.map((hit) => hit._id));
-    }
-    getAndSetAllUUIDs();
-  }, [queryBody]);
-
-  return { allResultsUUIDs, setQueryBodyAndReturnBody };
+  return { queryBody, setQueryBodyAndReturnBody };
 }
 
 const query = new CustomQuery({
@@ -68,7 +50,8 @@ function useSearch() {
 
   const defaultSortConfig = availableFields.mapped_last_modified_timestamp;
 
-  const { allResultsUUIDs, setQueryBodyAndReturnBody } = useAllResultsUUIDs();
+  const { queryBody, setQueryBodyAndReturnBody } = useQueryBody();
+
   const config = useMemo(() => {
     return {
       host: elasticsearchEndpoint,
@@ -105,11 +88,14 @@ function useSearch() {
   ]);
 
   const variables = useSearchkitVariables();
-  const { results } = useSearchkitSDK({
+  const { results, allResultsUUIDs } = useSearchkitSDK({
     config,
     variables,
     filters: defaultFilterValues,
     defaultSort: `${defaultSortConfig.field}.desc`,
+    queryBody,
+    elasticsearchEndpoint,
+    groupsToken,
   });
 
   return { results, entityType, allResultsUUIDs };