+++ context/app/static/js/components/workspaces/CreateWorkspaceDialog/CreateWorkspaceDialog.jsx
@@ -1,0 +1,42 @@
+import React from 'react';
+import Button from '@material-ui/core/Button';
+
+import CreateWorkspaceInput from 'js/components/workspaces/CreateWorkspaceInput';
+import DialogModal from 'js/shared-styles/DialogModal';
+import { useCreateWorkspace } from './hooks';
+
+function CreateWorkspaceDialog({ handleCreateWorkspace, buttonComponent: ButtonComponent, defaultName, ...rest }) {
+  const { dialogIsOpen, setDialogIsOpen, handleSubmit, handleClose, control, errors, onSubmit } = useCreateWorkspace({
+    handleCreateWorkspace,
+    defaultName,
+  });
+  return (
+    <>
+      <ButtonComponent onClick={() => setDialogIsOpen(true)} {...rest} />
+      <DialogModal
+        title="Create New Workspace"
+        isOpen={dialogIsOpen}
+        handleClose={handleClose}
+        maxWidth="md"
+        content={
+          <>
+            <form onSubmit={handleSubmit(onSubmit)}>
+              <CreateWorkspaceInput control={control} name="name" errors={errors} />
+              <input type="submit" id="create-workspace-input" hidden />
+            </form>
+          </>
+        }
+        actions={
+          <>
+            <Button onClick={handleClose}>Cancel</Button>
+            <label htmlFor="create-workspace-input">
+              <Button component="span">Submit</Button>
+            </label>
+          </>
+        }
+      />
+    </>
+  );
+}
+
+export default CreateWorkspaceDialog;