+++ context/app/static/js/components/workspaces/CreateWorkspaceDialog/index.js
@@ -1,0 +1,3 @@
+import CreateWorkspaceDialog from './CreateWorkspaceDialog';
+
+export default CreateWorkspaceDialog;