+++ context/app/static/js/components/workspaces/CreateWorkspaceInput/index.js
@@ -1,0 +1,3 @@
+import CreateWorkspaceInput from './CreateWorkspaceInput';
+
+export default CreateWorkspaceInput;