+++ context/app/static/js/components/workspaces/utils.js
@@ -23,13 +23,10 @@ async function createEmptyWorkspace({ workspacesEndpoint, workspacesToken, works
 }
 
 async function stopJob({ jobId, workspacesEndpoint, workspacesToken }) {
-  const response = await fetch(`${workspacesEndpoint}/jobs/${jobId}/stop/`, {
+  return fetch(`${workspacesEndpoint}/jobs/${jobId}/stop/`, {
     method: 'PUT',
     headers: getWorkspacesApiHeaders(workspacesToken),
-  });
-  if (!response.ok) {
-    throw Error(`Job stop for job #${jobId} failed`);
-  }
+  }).catch(() => console.error(`Job stop for job #${jobId} failed`));
 }
 
 async function deleteWorkspace({ workspaceId, workspacesEndpoint, workspacesToken }) {
@@ -48,28 +45,34 @@ async function stopJobs({ workspaceId, workspacesEndpoint, workspacesToken }) {
   const jobsResponse = await fetch(`${workspacesEndpoint}/jobs`, { headers });
   const jobsResults = await jobsResponse.json();
   const { jobs } = jobsResults.data;
-  jobs.forEach((job) => {
-    if (String(job.workspace_id) === String(workspaceId)) {
-      stopJob({ jobId: job.id, workspacesEndpoint, workspacesToken });
-    }
-  });
+
+  const activeWorkspaceJobs = jobs.filter(
+    (job) => String(job.workspace_id) === String(workspaceId) && ['running', 'pending'].includes(job.status),
+  );
+
+  await Promise.all(
+    activeWorkspaceJobs.map((job) => {
+      return stopJob({ jobId: job.id, workspacesEndpoint, workspacesToken });
+    }),
+  );
 }
 
 async function startJob({ workspaceId, workspacesEndpoint, workspacesToken, setMessage, setDead }) {
-  const startResponse = await fetch(`${workspacesEndpoint}/workspaces/${workspaceId}/start`, {
-    method: 'PUT',
-    headers: getWorkspacesApiHeaders(workspacesToken),
-    body: JSON.stringify({
-      job_type: 'JupyterLabJob',
-      job_details: {},
-    }),
-  });
-
-  if (!startResponse.ok) {
-    setDead(true);
-  }
-  const start = await startResponse.json();
-  setMessage(start.message);
+  fetch(`${workspacesEndpoint}/job_types`)
+    .then((response) => response.json())
+    .then(({ data: { job_types } }) =>
+      fetch(`${workspacesEndpoint}/workspaces/${workspaceId}/start`, {
+        method: 'PUT',
+        headers: getWorkspacesApiHeaders(workspacesToken),
+        body: JSON.stringify({
+          job_type: job_types.jupyter_lab.id,
+          job_details: {},
+        }),
+      }),
+    )
+    .then((response) => response.json())
+    .then(({ start }) => setMessage(start.message))
+    .catch(() => setDead(true));
 }
 
 function getNotebookPath(workspace) {
@@ -116,6 +119,7 @@ function condenseJobs(jobs) {
     const displayStatus = {
       pending: ACTIVATING,
       running: ACTIVE,
+      stopping: INACTIVE,
       complete: INACTIVE,
       failed: INACTIVE,
     }[status];