+++ context/app/markdown/CHANGELOG.md
@@ -1,3 +1,38 @@
+## v0.60.0 - 2022-12-20
+
+- Handle change in elasticsearch document from specimen_type and mapped_specimen_type to sample_category.
+- Added 4 more HRA previews
+- Updated the CCF Portal to the HRA Portal and it's new home, humanatlas.io
+
+
+## v0.59.0 - 2022-12-13
+
+- Correct protein measurement labels to abundance in cells UI.
+- Reset gene symbols/protein options when query type changes in autocomplete input.
+- Update new search ui to only retain selections of hits included in the new results.
+- Updates to support HRA Release 4 (v1.3)
+- Enable users to launch workspace notebooks from dataset selections in new search ui.
+- Enable users to select search results in new search ui.
+- Add disclaimer to top of cells page.
+
+
+## v0.57.0 - 2022-11-09
+
+- Fix path in ls command cell in workspace visualization notebook.
+- Updated the make_organ_dir.py logic to capture more HRA Reference Organs for display
+
+
+## v0.56.1 - 2022-11-01
+
+- Set workspace notebook path as URL param and use to create jupyter URL.
+- Remove leading slash in workspace symlink name path.
+- Update workspaces list after workspace create and delete so users no longer have to refresh the page for updates.
+- Handle stopping workspace jobs.
+- Launch dataset workspace notebook in new tab.
+- Add form to allow users to specify a new workspace name.
+- Pass job_types_id to workspace job start request.
+
+
 ## v0.56.0 - 2022-10-21
 
 - Add additional email for Phil Blood to workspaces allow list.